feat: Add confirmation steps to payroll and reset commands
This commit introduces confirmation steps for the `/payroll` and `/reset` administrative commands to prevent accidental data loss.

The `/payroll` command now includes a required `confirm: True` boolean parameter. The command will only proceed if this parameter is provided and set to `True`.

The `/reset` command now has a two-step confirmation process. It first requires a `confirm: True` parameter. If confirmed, it then presents the administrator with a secondary confirmation message containing "Confirm" and "Cancel" buttons. The irreversible data reset will only occur if the "Confirm" button is clicked.

A reusable `ConfirmView` class has been added to the `views/` directory to handle the button-based confirmation logic.

All relevant tests have been updated to account for these new confirmation flows, including new targeted tests for these features and smoke tests for other commands to ensure continued coverage. The test suite now passes.
